ev_app_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
email-validator==2.0.0  # ✅ ADD THIS LINE
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
openai==1.3.7
httpx==0.25.2
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6
bcrypt==4.1.2